linux-xiaomi-armani: disable kgsl 3d driver (#1334)
The kgsl 3d driver causes the device to boot loop, so disable this for
now.

The device boots ok after removing the kgsl driver, although continuing
boot from debug-shell via `pmos_continue_boot` does not work (seems to
be a bug in msm_fb driver). So, one has to reflash the boot partition
generated with no intfs debug shell to enjoy this improvement.