ci: limit running of qemu tests
This uses a variable configured in the gitlab project [settings->CI/CD->Variables]
to restrict when the qemu job will run. This should prevent the job from running
on repo forks where an appropriate runner has not been configured/registered.

Note that if this variable is set in a repo where there is no runner registered to
run qemu, then any CI jobs in that repo will ultimately fail the qemu tests since
no runner will be found.