test_newapkbuild: disable half test, upstream bug
Half of this test is failing, because of a regression in newapkbuild.

This patch will fix it:
https://github.com/alpinelinux/abuild/pull/75

Until then, let's disable this test, so our testsuite runs through
again. If somebody is curious on how I've developed and tested this
patch with pmbootstrap, I've added a guide to the wiki:
https://wiki.postmarketos.org/wiki/Hacking_on_abuild