gitlab-ci.yml: allow 'git config' cmds to fail
Two "su pmos -c 'git config --global user....'" commands were added to
the before_script in .gitlab-ci.yml. They work fine with the regular
gitlab runners, but fail with the custom gitlab runner that runs the
qemu test. Simply allow the command to fail, because it isn't needed on
the custom gitlab runner, as it doesn't run related tests.

Fixes: cf91601c817e ("gitlab-ci.yml: set git user/email (!1848)")